backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
anthropic>=0.25.0
pandas>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
pdfplumber>=0.10.0
aiofiles>=23.0.0
openpyxl>=3.1.0
